+++ gh-badges/lib/badge-cli.js
