+++ gh-badges/lib/colorscheme.json
