+++ gh-badges/lib/defaults.js
