+++ gh-badges/lib/make-badge.js
