+++ gh-badges/lib/make-badge.spec.js
