+++ gh-badges/lib/svg-to-img.js
