+++ gh-badges/lib/svg-to-img.spec.js
