+++ gh-badges/lib/text-measurer.js
