+++ gh-badges/templates/default-template.json
