+++ lib/badge-data.js
@@ -4,7 +4,7 @@ const isCSSColor = require('is-css-color')
 const logos = require('./load-logos')()
 const simpleIcons = require('./load-simple-icons')()
 const { svg2base64, isDataUri } = require('./logo-helper')
-const colorschemes = require('./colorscheme.json')
+const colorschemes = require('../gh-badges/lib/colorscheme.json')
 
 function toArray(val) {
   if (val === undefined) {
