+++ lib/refactoring-cli.js
@@ -3,7 +3,7 @@
 const chalk = require('chalk')
 const mapValues = require('lodash.mapvalues')
 
-const colorscheme = require('../lib/colorscheme.json')
+const colorscheme = require('../gh-badges/lib/colorscheme.json')
 const colorsMap = mapValues(colorscheme, 'colorB')
 const { floorCount } = require('./color-formatters')
 const { loadServiceClasses } = require('../services')
