+++ lib/request-handler.js
@@ -5,7 +5,7 @@ const domain = require('domain')
 const request = require('request')
 const { makeBadgeData: getBadgeData } = require('./badge-data')
 const log = require('./log')
-const LruCache = require('./lru-cache')
+const LruCache = require('../gh-badges/lib/lru-cache')
 const analytics = require('./analytics')
 const { makeSend } = require('./result-sender')
 const queryString = require('query-string')
