+++ lib/request-handler.spec.js
@@ -13,7 +13,7 @@ const {
   _requestCache,
   getBadgeMaxAge,
 } = require('./request-handler')
-const testHelpers = require('./make-badge-test-helpers')
+const testHelpers = require('../gh-badges/lib/make-badge-test-helpers')
 
 const handleRequest = makeHandleRequestFn(testHelpers.makeBadge())
