+++ lib/result-sender.js
@@ -2,7 +2,7 @@
 
 const stream = require('stream')
 const log = require('./log')
-const svg2img = require('./svg-to-img')
+const svg2img = require('../gh-badges/lib/svg-to-img')
 
 function streamFromString(str) {
   const newStream = new stream.Readable()
