+++ lib/server-config.js
@@ -5,7 +5,7 @@
 
 const url = require('url')
 const envFlag = require('node-env-flag')
-const defaults = require('./defaults')
+const defaults = require('../gh-badges/lib/defaults')
 
 function envArray(envVar, defaultValue, delimiter) {
   delimiter = delimiter || ','
