+++ now.json
@@ -7,6 +7,7 @@
     "package-lock.json",
     "build/",
     "frontend/",
+    "gh-badges/",
     "lib/",
     "logo/",
     "pages/",
