+++ services/base-svg-scraping.spec.js
@@ -5,7 +5,7 @@ const { expect } = require('chai')
 const sinon = require('sinon')
 const Joi = require('joi')
 const { makeBadgeData } = require('../lib/badge-data')
-const testHelpers = require('../lib/make-badge-test-helpers')
+const testHelpers = require('../gh-badges/lib/make-badge-test-helpers')
 const BaseSvgScrapingService = require('./base-svg-scraping')
 
 chai.use(require('chai-as-promised'))
